requirements/base.txt
Django==6.0.1
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.1.0
pandasai>=3.0.0
anthropic>=0.18.0
pandas>=2.0.0
openpyxl>=3.1.0
plotly>=5.18.0
python-magic-bin>=0.4.14
gunicorn==21.2.0
